add the skeleton component for all tweet cards
