+++ src/components/coursespage/Testimonial/index.jsx
@@ -1,6 +1,5 @@
-import Carousel from 'better-react-carousel';
 import clsx from 'clsx';
-import { Tweet } from 'react-twitter-widgets';
+import SkeletonCard from 'components/SkeletonCard';
 
 const Testimonials = () => {
   return (
@@ -18,31 +17,11 @@ const Testimonials = () => {
       >
         <h2 className='h1'>Testimonials</h2>
         <hr className='styled-hr' />
-        <Carousel
-          cols={3}
-          rows={1}
-          gap={20}
-          autoplay={2500}
-          dotColorActive='rgb(89,66,233)'
-          showDots
-          hideArrow
-          loop
-        >
-          {tweetId.map((id) => (
-            <Carousel.Item
-              style={{
-                width: '99%',
-              }}
-              key={id}
-            >
-              <Tweet
-                tweetId={id}
-                options={{ theme: 'dark', conversation: 'none' }}
-              />
-            </Carousel.Item>
-          ))}
-        </Carousel>
       </div>
+      <SkeletonCard
+        tweetId={tweetId}
+        options={{ theme: 'dark', conversation: 'none' }}
+      />
     </section>
   );
 };
