+++ src/components/hashnode/HowTakePart/index.jsx
@@ -1,7 +1,6 @@
-import Carousel from 'better-react-carousel';
 import ColumnSection from 'components/layout/ColumnSection';
 import ListItem from 'components/Listitem';
-import { Tweet } from 'react-twitter-widgets';
+import SkeletonCard from 'components/SkeletonCard';
 
 const HowTakePart = ({ background = 'static' }) => {
   return (
@@ -45,28 +44,8 @@ const HowTakePart = ({ background = 'static' }) => {
       >
         <h2 className='h1'>Previous winners</h2>
         <hr className='styled-hr' />
-        <Carousel
-          cols={3}
-          rows={1}
-          gap={20}
-          autoplay={2500}
-          dotColorActive='rgb(89,66,233)'
-          showDots
-          hideArrow
-          loop
-        >
-          {tweetId.map((id) => (
-            <Carousel.Item
-              style={{
-                width: '99%',
-              }}
-              key={id}
-            >
-              <Tweet tweetId={id} options={{ theme: 'dark' }} />
-            </Carousel.Item>
-          ))}
-        </Carousel>
       </div>
+      <SkeletonCard tweetId={tweetId} options={{ theme: 'dark' }} />
     </div>
   );
 };
