+++ src/components/home/events/index.jsx
@@ -1,28 +1,9 @@
 import BannerImage from 'assets/home/latestevents.png';
-import Carousel from 'better-react-carousel';
 import ColumnSection from 'components/layout/ColumnSection';
 import ListItem from 'components/Listitem';
-import { useEffect, useState } from 'react';
-import Skeleton from 'react-loading-skeleton';
-import { eagerLoadTwitterLibrary, Tweet } from 'react-twitter-widgets';
+import SkeletonCard from 'components/SkeletonCard';
 
-function InlineWrapperWithMargin({ children }) {
-  return (
-    <span
-      style={{
-        marginRight: '1.2rem',
-        marginLeft: '1.2rem',
-      }}
-    >
-      {children}
-    </span>
-  );
-}
 const Events = () => {
-  const [hideTweetPost, setHideTweetPost] = useState(true);
-  useEffect(() => {
-    eagerLoadTwitterLibrary();
-  }, []);
   return (
     <section>
       <ColumnSection
@@ -63,53 +44,10 @@ const Events = () => {
           </span>
         </ListItem>
       </ColumnSection>
-      {hideTweetPost ? (
-        <div className='layout'>
-          <Skeleton
-            count={3}
-            wrapper={InlineWrapperWithMargin}
-            inline
-            width='30%'
-            height='71vh'
-            baseColor='#15202b'
-            highlightColor='#6555cc'
-            borderRadius='0.5rem'
-          />
-        </div>
-      ) : null}
-      <div
-        className='layout'
-        style={{
-          paddingBottom: '5rem',
-          opacity: hideTweetPost ? '0' : '1',
-        }}
-      >
-        <Carousel
-          cols={3}
-          rows={1}
-          gap={20}
-          autoplay={2500}
-          dotColorActive='rgb(89,66,233)'
-          showDots
-          hideArrow
-          loop
-        >
-          {tweetId.map((id) => (
-            <Carousel.Item
-              style={{
-                width: '99%',
-              }}
-              key={id}
-            >
-              <Tweet
-                tweetId={id}
-                options={{ theme: 'dark' }}
-                onLoad={() => setHideTweetPost(false)}
-              />
-            </Carousel.Item>
-          ))}
-        </Carousel>
-      </div>
+      <SkeletonCard
+        tweetId={tweetId}
+        options={{ theme: 'dark', conversation: 'none' }}
+      />
     </section>
   );
 };
